Conflicts:
	src/main/aerospike/as_operations.c
	src/test/aerospike_cdt/cdt_basics.c

+++ src/test/aerospike_cdt/cdt_basics.c
@@ -390,7 +390,7 @@ as_testlist_insert(as_testlist *tlist, int64_t index, as_val *val)
 	as_operations ops;
 	as_operations_inita(&ops, 1);
 
-	uint64_t uindex = index2uindex(tlist, index);
+	uint32_t uindex = (uint32_t)index2uindex(tlist, index);
 	as_val_reserve(val);
 	as_arraylist_insert(&tlist->arraylist, uindex, val);
 
@@ -405,7 +405,7 @@ as_testlist_insert_list(as_testlist *tlist, int64_t index, as_arraylist *list)
 	as_operations ops;
 	as_operations_inita(&ops, 1);
 
-	uint64_t uindex = index2uindex(tlist, index);
+	uint32_t uindex = (uint32_t)index2uindex(tlist, index);
 	uint32_t size = as_arraylist_size(list);
 	for (int i = size - 1; i >= 0; i--) {
 		as_val *val = as_arraylist_get(list, i);
@@ -424,7 +424,7 @@ as_testlist_set(as_testlist *tlist, int64_t index, as_val *val)
 	as_operations ops;
 	as_operations_inita(&ops, 1);
 
-	uint64_t uindex = index2uindex(tlist, index);
+	uint32_t uindex = (uint32_t)index2uindex(tlist, index);
 	as_val_reserve(val);
 	as_arraylist_set(&tlist->arraylist, uindex, val);
 
@@ -439,8 +439,9 @@ as_testlist_trim(as_testlist *tlist, int64_t index, uint64_t count)
 	as_operations ops;
 	as_operations_inita(&ops, 1);
 
-	uint64_t uindex = index2uindex(tlist, index);
-	as_arraylist_trim(&tlist->arraylist, uindex + count);
+	uint32_t uindex = (uint32_t)index2uindex(tlist, index);
+	as_arraylist_trim(&tlist->arraylist, uindex + (uint32_t)count);
+
 	for (int64_t i = 0; i < index; i++) {
 		as_arraylist_remove(&tlist->arraylist, 0);
 	}
